cookus-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.0
mysql-connector-python==9.2.0
pydantic==2.10.3
PyJWT==2.*
python-jose[cryptography]==3.*
fastapi-mail==1.4.1